chore: extracted symlink git client
