+++ cli/commands/run/download_source.go
@@ -3,7 +3,6 @@ package run
 import (
 	"context"
 	"fmt"
-	"net/url"
 	"os"
 	"path/filepath"
 	"strings"
@@ -247,7 +246,6 @@ func UpdateGetters(terragruntOptions *options.TerragruntOptions, terragruntConfi
 
 		for getterName, getterValue := range getter.Getters {
 			if getterName == "file" {
-
 				var includeInCopy, excludeFromCopy []string
 
 				if terragruntConfig.Terraform != nil && terragruntConfig.Terraform.IncludeInCopy != nil {
@@ -296,38 +294,6 @@ func preserveSymlinksOption() getter.ClientOption {
 
 		return nil
 	}
-}
-
-// symlinkPreservingGitGetter wraps the original git getter to preserve symlink settings
-type symlinkPreservingGitGetter struct {
-	original getter.Getter
-	client   *getter.Client
-}
-
-func (g *symlinkPreservingGitGetter) Get(dst string, u *url.URL) error {
-	// Store original DisableSymlinks setting
-	originalDisableSymlinks := g.client.DisableSymlinks
-
-	// Call the original getter
-	err := g.original.Get(dst, u)
-
-	// Restore the original DisableSymlinks setting
-	g.client.DisableSymlinks = originalDisableSymlinks
-
-	return err
-}
-
-func (g *symlinkPreservingGitGetter) GetFile(dst string, u *url.URL) error {
-	return g.original.GetFile(dst, u)
-}
-
-func (g *symlinkPreservingGitGetter) ClientMode(u *url.URL) (getter.ClientMode, error) {
-	return g.original.ClientMode(u)
-}
-
-func (g *symlinkPreservingGitGetter) SetClient(c *getter.Client) {
-	g.client = c
-	g.original.SetClient(c)
 }
 
 // Download the code from the Canonical Source URL into the Download Folder using the go-getter library
